requirements.txt
aiohttp==3.6.2
async-timeout==3.0.1
attrs==20.2.0
blis==0.4.1
cachetools==4.1.1
catalogue==1.0.0
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
cymem==2.0.3
DateTime==4.3
Flask==1.1.2
google-api-core==1.22.3
google-auth==1.22.0
google-cloud-language==1.3.0
googleapis-common-protos==1.52.0
grpcio==1.32.0
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
jmespath==0.10.0
jsonify==0.5
MarkupSafe==1.1.1
multidict==4.7.6
murmurhash==1.0.2
numpy==1.19.2
oauthlib==3.1.0
pandas==1.1.2
plac==1.1.3
preshed==3.0.2
protobuf==3.13.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
PySocks==1.7.1
python-dateutil==2.8.1
python-dotenv==0.14.0
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
s3transfer==0.3.3
six==1.15.0
srsly==1.0.2
thinc==7.4.1
tqdm==4.50.0
tweepy==3.9.0
urllib3==1.25.10
wasabi==0.8.0
Werkzeug==1.0.1
yarl==1.6.0
zope.interface==5.1.1
